cmd: prevent ongoing requests being canceled by deadline exceeded
