+++ cmd/client_meta.go
@@ -20,8 +20,7 @@ func newDeleteMetaCommand() *cobra.Command {
 		Use:   "delete",
 		Short: "Delete all meta data in etcd, confirm that you know what this command will do and use it at your own risk",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			ctx, cancel := contextTimeout()
-			defer cancel()
+			ctx := defaultContext
 			err := cdcEtcdCli.ClearAllCDCInfo(ctx)
 			if err == nil {
 				cmd.Println("already truncate all meta in etcd!")
